+++ R/trans.r
@@ -14,7 +14,7 @@
 #'   inverse of the transformation
 #' @param breaks default breaks function for this transformation. The breaks
 #'   function is applied to the raw data.
-#' @param minor.breaks default minor breaks function for this transformation.
+#' @param minor_breaks default minor breaks function for this transformation.
 #' @param format default format for this transformation. The format is applied
 #'   to breaks generated to the raw data.
 #' @param domain domain, as numeric vector of length 2, over which
@@ -23,8 +23,8 @@
 #' @export trans_new is.trans
 #' @aliases trans_new trans is.trans
 trans_new <- function(name, transform, inverse, breaks = extended_breaks(),
-                      minor.breaks = minor_breaks(), format = format_format(),
-                      domain = c(-Inf, Inf)) {
+                      minor_breaks = even_minor_breaks(),
+                      format = format_format(), domain = c(-Inf, Inf)) {
   if (is.character(transform)) transform <- match.fun(transform)
   if (is.character(inverse)) inverse <- match.fun(inverse)
 
@@ -34,7 +34,7 @@ trans_new <- function(name, transform, inverse, breaks = extended_breaks(),
       transform = transform,
       inverse = inverse,
       breaks = breaks,
-      minor_breaks = minor_breaks,
+      minor_breaks = even_minor_breaks,
       format = format,
       domain = domain),
     class = "trans"