+++ NEWS.md
@@ -1,6 +1,6 @@
 # scales 0.4.1.9001
 
-* New function `calculate_minor_breaks()` calculates minor breaks as a property
+* New function `even_minor_breaks()` calculates minor breaks as a property
   of the transformation (@karawoo).
 
 * Adds `viridis_pal()` for creating palettes with color maps from the